scripts/n8n-compact/requirements.txt
python-dotenv==1.0.1p